starts to work
naming and unit backref needed

+++ src/mopeka.h
@@ -2,27 +2,27 @@
 #include <stdint.h>
 
 typedef enum {
-  // Gen2/Standard
-  STD = 2,
-  XL = 3,
-  BMPRO_STD = 70,
+    // Gen2/Standard
+    STD = 2,
+    XL = 3,
+    BMPRO_STD = 70,
 
-  // Pro Family IDs offset by 256 to separate from standard sensors
-  UNKNOWN = 0 | 0x100, // If PRO is reporting this, it likely indicates an
-                       // obscure hardware issue
-  VERTRAX_STANDARD = 1 | 0x100,
-  VERTRAX_BULK = 2 | 0x100,
-  PRO_MOPEKA = 3 | 0x100, // Mopeka PRO LPG sensor
-  TOPDOWN = 4 | 0x100, // This will is currently both the top-down water and air
-                       // sensor, but will likely break it out in the future
-  PRO_H2O = 5 | 0x100, // This is the PRO sensor, but with hwid for doing
-                       // bottom-up assumed in a water tank
-  PRO_LIPPERT_LPG = 6 | 0x100,     // Lippert private labeled LPG sensor
-  PRO_DOMETIC_LPG = 7 | 0x100,     // Dometic private labeled LPG sensor
-  PRO_PLUS_BLE_LPG = 8 | 0x100,    // Mopeka PRO+ LPG sensor, BLE boosted
-  PRO_PLUS_CELL_LPG = 9 | 0x100,   // Mopeka PRO+ LPG sensor, Cellular
-  PRO_PLUS_BLE_TD40 = 10 | 0x100,  // Mopeka PRO+ TD40 LPG sensor, BLE booster
-  PRO_PLUS_CELL_TD40 = 11 | 0x100, // Mopeka PRO+ TD40 LPG sensor, Cellular
+    // Pro Family IDs offset by 256 to separate from standard sensors
+    UNKNOWN = 0 | 0x100, // If PRO is reporting this, it likely indicates an
+    // obscure hardware issue
+    VERTRAX_STANDARD = 1 | 0x100,
+    VERTRAX_BULK = 2 | 0x100,
+    PRO_MOPEKA = 3 | 0x100, // Mopeka PRO LPG sensor
+    TOPDOWN = 4 | 0x100, // This will is currently both the top-down water and air
+    // sensor, but will likely break it out in the future
+    PRO_H2O = 5 | 0x100, // This is the PRO sensor, but with hwid for doing
+    // bottom-up assumed in a water tank
+    PRO_LIPPERT_LPG = 6 | 0x100,     // Lippert private labeled LPG sensor
+    PRO_DOMETIC_LPG = 7 | 0x100,     // Dometic private labeled LPG sensor
+    PRO_PLUS_BLE_LPG = 8 | 0x100,    // Mopeka PRO+ LPG sensor, BLE boosted
+    PRO_PLUS_CELL_LPG = 9 | 0x100,   // Mopeka PRO+ LPG sensor, Cellular
+    PRO_PLUS_BLE_TD40 = 10 | 0x100,  // Mopeka PRO+ TD40 LPG sensor, BLE booster
+    PRO_PLUS_CELL_TD40 = 11 | 0x100, // Mopeka PRO+ TD40 LPG sensor, Cellular
 } mopekaHwOID;
 
 typedef enum { MOPEKA_CC2540, MOPEKA_NRF52, MOPEKA_GW } mopekaAdvertID;
@@ -43,25 +43,22 @@ typedef enum { MOPEKA_CC2540, MOPEKA_NRF52, MOPEKA_GW } mopekaAdvertID;
 
 // // propane/butane corr:  https://github.com/Albertworker1004/Mopeka_React-Native/blob/main/Tank%20Check/src/lib/sensors/tankcheck.ts#L574
 
-
 typedef struct {
 
-  float temperature;
-  float level;
+    int16_t temperature;
+    int16_t level;
     float battery;
-  int32_t raw_level;
-  int8_t raw_temp;
-  int8_t rssi;
-  uint8_t hw_id;
-  uint8_t qualityStars;
+    int32_t raw_level;
+    int8_t raw_temp;
+    int8_t rssi;
+    uint8_t hw_id;
+    uint8_t qualityStars;
 
-  int8_t acceloX;
-  int8_t acceloY;
+    int8_t acceloX;
+    int8_t acceloY;
 
-  bool syncPressed;
-  bool slowUpdateRate;
-  bool corrupted;
+    bool syncPressed;
+    bool slowUpdateRate;
+    bool corrupted;
 
 } mopekaAd_t;
-
-
